+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     * `--no-live-config-reload`
     * `--dimensions`
     * `--position`
+- `live-shader-reload` feature
 
 ## 0.6.0